+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/mania/widgets/ManiaWidgetVPSTableAlxController.java
@@ -6,11 +6,14 @@ import de.mephisto.vpin.connectors.mania.model.ManiaVpsTable;
 import de.mephisto.vpin.connectors.vps.model.VpsTable;
 import de.mephisto.vpin.connectors.vps.model.VpsTableVersion;
 import de.mephisto.vpin.ui.Studio;
+import de.mephisto.vpin.ui.mania.ManiaController;
 import de.mephisto.vpin.ui.mania.TarcisioWheelsDB;
 import de.mephisto.vpin.ui.tournaments.VpsTableContainer;
 import de.mephisto.vpin.ui.tournaments.VpsVersionContainer;
 import javafx.application.Platform;
 import javafx.beans.property.SimpleObjectProperty;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
@@ -18,11 +21,14 @@ import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.geometry.Pos;
 import javafx.scene.Parent;
+import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+import javafx.scene.input.MouseButton;
+import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.StackPane;
@@ -63,8 +69,14 @@ public class ManiaWidgetVPSTableAlxController extends WidgetController implement
   @FXML
   private StackPane tableStack;
 
+  @FXML
+  private Button tableStatsBtn;
+
   private Parent loadingOverlay;
+
   private List<ManiaVpsTable> vpsTables;
+
+  private ManiaController maniaController;
 
   // Add a public no-args constructor
   public ManiaWidgetVPSTableAlxController() {
@@ -75,9 +87,30 @@ public class ManiaWidgetVPSTableAlxController extends WidgetController implement
     refresh();
   }
 
+  @FXML
+  private void onTableStats() {
+    ManiaVpsTable selectedItem = tableView.getSelectionModel().getSelectedItem();
+    if (selectedItem != null) {
+      VpsTable tableById = Studio.client.getVpsService().getTableById(selectedItem.getVpsTableId());
+      if (tableById != null) {
+        maniaController.selectVpsTable(tableById);
+      }
+    }
+  }
+
+  @FXML
+  private void onTableMouseClicked(MouseEvent mouseEvent) {
+    if (mouseEvent.getButton().equals(MouseButton.PRIMARY)) {
+      if (mouseEvent.getClickCount() == 2) {
+        onTableStats();
+      }
+    }
+  }
+
   @Override
   public void initialize(URL url, ResourceBundle resourceBundle) {
     tableView.setPlaceholder(new Label("No score submitted yet."));
+    tableStatsBtn.setDisable(true);
 
     columnRank.setCellValueFactory(cellData -> {
       ManiaVpsTable value = cellData.getValue();
@@ -146,6 +179,13 @@ public class ManiaWidgetVPSTableAlxController extends WidgetController implement
     catch (IOException e) {
       LOG.error("Failed to load loading overlay: " + e.getMessage());
     }
+
+    tableView.getSelectionModel().selectedItemProperty().addListener(new ChangeListener<ManiaVpsTable>() {
+      @Override
+      public void changed(ObservableValue<? extends ManiaVpsTable> observable, ManiaVpsTable oldValue, ManiaVpsTable newValue) {
+        tableStatsBtn.setDisable(newValue == null);
+      }
+    });
   }
 
   public void refresh() {
@@ -159,4 +199,9 @@ public class ManiaWidgetVPSTableAlxController extends WidgetController implement
       });
     }).start();
   }
+
+  public void setManiaController(ManiaController maniaController) {
+    this.maniaController = maniaController;
+  }
+
 }