+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ## Changes
 
+- **VPin Mania / Table Statistics**: Added open button/open on double-click so that the table details are shown.
 - **Preferences / Highscore Settings**: Added "NVRam Synchronization" button to re-download all available resetted nvram files. Usually you won't need this since new nvrams are downloaded on server startup, but this way you can update them e.g. in case the volume was adjusted too.
 - **System Manager**: Releases are now selectable via their actual name, not the tagged version anymore. Especially VPX has more meaningful names there.
 - **System Manager**: Added textarea that shows the release notes of the selected release.